+++ app/src/main/java/solutions/silly/wearnetatmo/model/Place.kt
@@ -1,7 +1,7 @@
 package solutions.silly.wearnetatmo.model
 
 data class Place(
-    val altitude: Long?,
+    val altitude: Int?,
     val city: String?,
     val country: String?,
     val timezone: String?,
